requirements-lock.txt
# Generated: pip freeze (conda env: oneclick)
adapter-transformers==4.0.0
albumentations==2.0.8
datasets==4.6.1
grad-cam==1.5.5
huggingface_hub==1.5.0
numpy==2.2.6
open_clip_torch==3.3.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
pillow==12.1.1
scipy==1.15.3
sentence-transformers==5.3.0
sklearn-crfsuite==0.3.6
strictyaml==1.7.3
timm==1.0.26
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.66.3
transformers==5.3.0
transformers-stream-generator==0.0.5
wandb==0.25.0
